+++ include/graphics/CircleRenderer.h
@@ -22,21 +22,21 @@ namespace Game3 {
 			void update(int width, int height) override;
 
 			void drawOnMap(const RenderOptions &, float cutoff = -1.f);
-			void drawOnScreen(const Color &color, double x, double y, double width, double height, float cutoff = -1.f, double angle = 0.);
+			void drawOnScreen(const Color &color, float x, float y, float width, float height, float cutoff = -1.f, float angle = 0.f);
 
 			inline auto getWidth()  const { return backbufferWidth;  }
 			inline auto getHeight() const { return backbufferHeight; }
 
 		private:
 			Canvas &canvas;
-			glm::dmat4 projection;
+			glm::mat4 projection;
 			GLuint quadVAO = 0;
 			int initializedTo = 0;
 
-			Lockable<std::map<int, std::vector<double>>> vertexMap;
+			Lockable<std::map<int, std::vector<float>>> vertexMap;
 
 			void initRenderData(int sides);
-			const std::vector<double> & getVertices(int sides, std::shared_lock<DefaultMutex> &);
+			const std::vector<float> & getVertices(int sides, std::shared_lock<DefaultMutex> &);
 			bool isInitialized() const;
 	};
 }
