+++ include/graphics/FluidRenderer.h
@@ -16,8 +16,8 @@ namespace Game3 {
 
 	class FluidRenderer {
 		public:
-			constexpr static double TEXTURE_SCALE = 2.;
-			constexpr static double TILE_TEXTURE_PADDING = 1. / 2048.;
+			constexpr static float TEXTURE_SCALE = 2.f;
+			constexpr static float TILE_TEXTURE_PADDING = 1.f / 2048.f;
 			std::atomic_int backbufferWidth = -1;
 			std::atomic_int backbufferHeight = -1;
 			std::atomic_bool isMissing = false;
@@ -31,7 +31,7 @@ namespace Game3 {
 			void reset();
 			void init();
 			void setup(TileProvider &);
-			void render(double divisor, double scale, double center_x, double center_y);
+			void render(float divisor, float scale, float center_x, float center_y);
 			bool reupload();
 			bool onBackbufferResized(int width, int height);
 			void setChunk(FluidChunk &, bool can_reupload = true);
